Extract windows
